+++ packages/client-node/src/connection/create_connection.ts
@@ -16,6 +16,7 @@ export interface CreateConnectionParams {
   http_agent: http.Agent | https.Agent | undefined
   set_basic_auth_header: boolean
   capture_enhanced_stack_trace: boolean
+  eagerly_destroy_stale_sockets?: boolean
 }
 
 /** A factory for easier mocking after Node.js 22.18 */
@@ -28,6 +29,7 @@ export class NodeConnectionFactory {
     http_agent,
     set_basic_auth_header,
     capture_enhanced_stack_trace,
+    eagerly_destroy_stale_sockets = false,
   }: CreateConnectionParams): NodeBaseConnection {
     if (http_agent !== undefined) {
       return new NodeCustomAgentConnection({
@@ -36,6 +38,7 @@ export class NodeConnectionFactory {
         capture_enhanced_stack_trace,
         keep_alive, // only used to enforce proper KeepAlive headers
         http_agent,
+        eagerly_destroy_stale_sockets,
       })
     }
     switch (connection_params.url.protocol) {
@@ -45,6 +48,7 @@ export class NodeConnectionFactory {
           set_basic_auth_header,
           capture_enhanced_stack_trace,
           keep_alive,
+          eagerly_destroy_stale_sockets,
         })
       case 'https:':
         return new NodeHttpsConnection({
@@ -53,6 +57,7 @@ export class NodeConnectionFactory {
           capture_enhanced_stack_trace,
           keep_alive,
           tls,
+          eagerly_destroy_stale_sockets,
         })
       default:
         throw new Error('Only HTTP and HTTPS protocols are supported')